refactor: move models to the separate files
