+++ ImageFeedApp/Services/ProfileImageService.swift
@@ -82,22 +82,3 @@ final class ProfileImageService {
     }
 
 }
-
-private struct UserResult: Codable {
-    enum CodingKeys: String, CodingKey {
-        case profileImage = "profile_image"
-    }
-
-    var profileImage: ProfileImage
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.profileImage = try container.decode(ProfileImage.self, forKey: .profileImage)
-    }
-}
-
-private struct ProfileImage: Codable {
-    let small: String
-    let medium: String
-    let large: String
-}
