+++ ImageFeedApp/Services/ProfileService.swift
@@ -33,7 +33,7 @@ final class ProfileService {
         let task = URLSession.shared.objectTask(for: request, completion: { [weak self] (result: Result<ProfileResult, Error>) in
             switch result {
             case .success(let response):
-                let profile = Profile(username: response.username, name: "\(response.firstName) \(response.lastName ?? "")", loginName: "@\(response.username)", bio: response.bio ?? "")
+                let profile = Profile(from: response)
                 completion(.success(profile))
             case .failure(let error):
                 print("[ProfileService] failed to make a request: \(error)")
@@ -74,34 +74,4 @@ final class ProfileService {
     func updateProfileDetails(profile: Profile) {
         self.profile = profile
     }
-}
-
-private struct ProfileResult: Decodable {
-    enum CodingKeys: String, CodingKey {
-        case username = "username"
-        case firstName = "first_name"
-        case lastName = "last_name"
-        case bio = "bio"
-    }
-
-    let username: String
-    let firstName: String
-    let lastName: String?
-    let bio: String?
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.username = try container.decode(String.self, forKey: .username)
-        self.firstName = try container.decode(String.self, forKey: .firstName)
-        self.lastName = try container.decode(String?.self, forKey: .lastName)
-        self.bio = try container.decode(String?.self, forKey: .bio)
-    }
-
-}
-
-struct Profile {
-    let username: String
-    let name: String
-    let loginName: String
-    let bio: String
 }
